feat(devcontainer.json): add devcontainer.json file to configure the development container for the project
The devcontainer.json file is added to the project. This file is used to configure the development container for the project. It specifies the name of the container, the base image to use, and the features to include in the container. In this case, the container is based on the "mcr.microsoft.com/devcontainers/base:jammy" image and includes features for Node.js, Jest, TypeScript, and npm. The VS Code extensions for ESLint, DeepScan, npm IntelliSense, path IntelliSense, Prettier, Sourcery, WakaTime, and code spell checker are also included.